requirements.txt
Django>=4.2
djangorestframework>=3.14

# Export functionality
openpyxl>=3.1.0

# File handling and uploads
Pillow>=10.0.0
